shifted wait_for_service to init function
