+++ ogc_ws/src/rff/src/rff_old.py
@@ -3,16 +3,16 @@
 import time
 import rospy
 import RuTOS
+import csv
 import threading
-import Adafruit_BBIO.GPIO as GPIO
 from mavros_msgs.srv import CommandBool
 from mavros_msgs.srv import SetMode
 from mavros_msgs.srv import WaypointPush
 from mavros_msgs.srv import WaypointSetCurrent
+from mavros_msgs.msg import WaypointList
 from mavros_msgs.msg import Waypoint
 from mavros_msgs.msg import State
-from std_msgs.msg import String
-from despatcher.msg import LinkMessage
+from pathlib import Path
 from waypoint import WP
 
 missionlist = []
@@ -20,28 +20,38 @@ missionlist = []
 class RFF:
 
 	def __init__(self):
-		# Set defaults
-		self.armStatus = True
-		self.hopStatus = False
-		self.Ack = False
-
 		# Check arm status of aircraft
 		rospy.Subscriber('/mavros/state', State, self.checkDisarm)
-		# Check hop status of aircraft
-		rospy.Subscriber('/ogc/to_rff', String, self.checkHop)
-
-		self.pub_to_aircraft = rospy.Publisher('ogc/from_sms', String, queue_size = 5)
+		# Path to missionlist.txt which contains all missions in a flight in order
+		path = str(waypointsfolder + "mission_list.txt")
+		f = open(path, "r")
+		for line in f:
+			# Ignores # comments
+			if line.startswith('#'):
+				continue
+			if line.startswith('Last update:'):
+				continue
+			missionlist.append(line.rstrip())
+		# Prints the missions for operator to check
+		rospy.loginfo("Missions for this flight:")
+		for i in missionlist:
+			rospy.loginfo(i)
+		# Assumes armed for safety
+		self.armStatus = True
 
 	def main(self):
-		if self.armStatus == True or self.hopStatus == False:
-			rospy.logwarn("Button disabled")
-			return
-		rospy.loginfo("Loading next set of waypoints.")
-		self.pub_to_aircraft.publish("i 0 10 mission next " + str(rospy.get_rostime().secs))
-		time.sleep(5)
-		if self.hopStatus == False or self.Ack == False:
-			rospy.loginfo("Unable to load")
-			return
+		rospy.loginfo("Loading waypoints.")
+		rospy.wait_for_service('mavros/mission/push', timeout=10)
+		wppush = rospy.ServiceProxy('mavros/mission/push', WaypointPush)
+		while not wppush(0, waypoints).success:
+			rospy.logerr("Failed to load waypoints. Trying again.")
+			time.sleep(3)
+		rospy.loginfo("Waypoints loaded.")
+		rospy.wait_for_service('mavros/mission/set_current', timeout=10)
+		wpset = rospy.ServiceProxy('mavros/mission/set_current', WaypointSetCurrent)
+		while not wpset(1).success:
+			rospy.logerr("Failed to set current waypoint. Trying again.")
+			time.sleep(3)
 		rospy.wait_for_service('mavros/set_mode', timeout=10)
 		mode = rospy.ServiceProxy('mavros/set_mode', SetMode)
 		while not mode(custom_mode = "AUTO").mode_sent:
@@ -57,18 +67,9 @@ class RFF:
 			time.sleep(3)
 		rospy.loginfo("THROTTLE ARMED")
 		self.armStatus = True
-		self.Ack = False
 
 	def checkDisarm(self, data):
 		self.armStatus = data.armed
-
-	def checkHop(self, data):
-		text = data.data.split()
-		if text[0] == "hop":
-			self.hopStatus = eval(text[1])
-		elif text[0] == "ack":
-			self.Ack = True
-
 
 class Button:
 
@@ -84,16 +85,8 @@ class Button:
 	def blink(self):
 		while True:
 			RuTOS.blink_button(self.ssh)
-			buzzer_pin = "P8_11"
-			GPIO.setup(buzzer_pin, GPIO.OUT)
-			for i in range(300):
-				GPIO.output(buzzer_pin, GPIO.HIGH)
-				time.sleep(0.0003)
-				GPIO.output(buzzer_pin, GPIO.LOW)
-				time.sleep(0.0003)
 			time.sleep(0.5)
 			if self.stopwarn:
-				GPIO.cleanup()
 				break
 
 	def press(self):
@@ -127,20 +120,24 @@ class Button:
 
 if __name__ == "__main__":
 	rospy.init_node('rff', anonymous=False, disable_signals=True)
+	waypointsfolder = rospy.get_param('~waypoint_folder', '/home/ubuntu/Waypoints/')
 	rff = RFF()
 
 	try:
-		while True:
-			# Do nothing while aircraft is not in hop mode
-			while not rff.hopStatus:
-				time.sleep(5)
-			rospy.loginfo("Plane is now in hop mode")
+		for i in range(len(missionlist)):
+			wpfile = str(waypointsfolder + missionlist[i])
+			wpread = WP()
+			waypoints = wpread.read(wpfile)
 			# Do nothing if aircraft is armed
 			while rff.armStatus:
 				time.sleep(5)
-			rospy.loginfo("Plane has been disarmed.")
+			rospy.loginfo("Plane has been disarmed. Starting button monitoring. ")
 			button = Button()
 			button.press()
+			rospy.loginfo(missionlist[i] + " started.")
+			waypoints = []
+		
+		rospy.loginfo("All missions finished")
 
 	except KeyboardInterrupt:
 		rospy.signal_shutdown("Shutting down")
